backend-files/websocket/requirements.txt
# requirements.txt
fastapi
uvicorn
psycopg2-binary
